requirements.txt
fastapi[all]
boto3
uvicorn
pyjwt
pydantic
sqlalchemy
pymysql
mysql-connector-python
passlib
python-jose[cryptography]
cryptography
python-dotenv
git+https://github.com/harshitsaini/srptools.git@keylance
Authlib
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
hypercorn
mysqlclient
